Moved Pdu to own package from canmatrix. No tests for this class
